add config loader
